backend/app/requirements.txt
Flask==2.1.2
google-cloud-storage==2.5.0